requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python>=1.0.3

requests~=2.31.0
beautifulsoup4~=4.12.2
lapps~=0.0.2
SPARQLWrapper~=2.0.0
